refactor(DEVING-24): 로그인 로직과 ui 분리 및 리펙토링
